+++ CHANGELOG.md
@@ -4,19 +4,29 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
+## [1.3.3] - unreleased
+
+### Added
+- Methods:
+    - `Page\Account\UserAddress::seeNumberOfShippingAddresses`
+### Fixed
+- Methods:
+    - `Page\Account\UserAddress::deleteShippingAddress`
+    - `Page\Component\Modal::confirmDeletion`
+
 ## [1.3.2] - 2020-04-21
 
 ### Added
 - Properties:
-    - `OxidEsales\Codeception\Page\Component\UserForm::$userNameField`
+    - `Page\Component\UserForm::$userNameField`
 - Methods:
-    - `OxidEsales\Codeception\Page\Component\UserForm::modifyUserName`
-    - `OxidEsales\Codeception\Step\Basket::addProductToBasketAndOpenUserCheckout`
-    - `OxidEsales\Codeception\Step\Basket::addProductToBasketAndOpenBasket`
+    - `Page\Component\UserForm::modifyUserName`
+    - `Step\Basket::addProductToBasketAndOpenUserCheckout`
+    - `Step\Basket::addProductToBasketAndOpenBasket`
 
 ### Deprecated
 - Methods:
-    - `OxidEsales\Codeception\Step\Basket::addProductToBasketAndOpen`
+    - `Step\Basket::addProductToBasketAndOpen`
 
 ## [1.3.1] - 2020-04-09
 
@@ -106,6 +116,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 ### Added
 - First version of the module introduced
 
+[1.3.3]: https://github.com/OXID-eSales/codeception-page-object/compare/v1.3.2..v1.3.3
+[1.3.2]: https://github.com/OXID-eSales/codeception-page-object/compare/v1.3.1..v1.3.2
 [1.3.1]: https://github.com/OXID-eSales/codeception-page-object/compare/v1.3.0..v1.3.1
 [1.3.0]: https://github.com/OXID-eSales/codeception-page-object/compare/v1.2.0..v1.3.0
 [1.2.0]: https://github.com/OXID-eSales/codeception-page-object/compare/v1.1.1..v1.2.0